refactor: create types
- Move create CloudinaryAsset type from gatsby-image to node-creation  module
