+++ packages/gatsby-transformer-cloudinary/node-creation/types.js
@@ -1,0 +1,7 @@
+exports.CloudinaryAssetType = `
+  type CloudinaryAsset implements Node @dontInfer  {
+    publicId: String!
+    cloudName: String!
+    version: String
+  }
+`;
